add error hanlder and modify some data type.

+++ docker/sql/postgres_init.sql
@@ -1,5 +1,5 @@
 CREATE TABLE IF NOT EXISTS public.account (
-    user_id INTEGER NOT NULL PRIMARY KEY,
+    user_id BIGINT NOT NULL PRIMARY KEY,
     available_balance NUMERIC(18,2) NOT NULL DEFAULT 0,
     reserved_balance NUMERIC(18,2) NOT NULL DEFAULT 0,
     updated_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP
@@ -8,8 +8,8 @@ CREATE TABLE IF NOT EXISTS public.account (
 CREATE TABLE IF NOT EXISTS public.transaction (
     transaction_id UUID NOT NULL DEFAULT gen_random_uuid(),
     nonce BIGINT NOT NULL,
-    from_account_id INTEGER NOT NULL,
-    to_account_id INTEGER NOT NULL,
+    from_account_id BIGINT NOT NULL,
+    to_account_id BIGINT NOT NULL,
     amount NUMERIC(18,2) NOT NULL,
     status INTEGER NOT NULL DEFAULT 0,
     created_at TIMESTAMP WITHOUT TIME ZONE NOT NULL DEFAULT CURRENT_TIMESTAMP,
